requirements.txt
flask==3.0.2
python-dotenv==1.0.1
twilio==8.13.0
openai==1.12.0
SQLAlchemy==2.0.27
python-decouple==3.8
gunicorn==21.2.0
requests==2.31.0 